requirements.txt

torch
torchvision
torchaudio

pillow
numpy

# Web app
flask

matplotlib

tqdm
